refactor(MetadataEditor): break 627-line component into hook + parts
- useMetadataEditor hook owns TVDB search, TVDB->form application, and save
  orchestration (show/episode upsert + extraction linking)
- TVDBSearchSection/TVDBResultItem, ShowMetadataForm, EpisodeMetadataForm split out
- New shared FormField/FormTextarea/FieldLabel primitives collapse ~13 repeated
  label+input blocks; modal now uses the shared ModalBase
- Component becomes a folder with index.tsx orchestrator; 3 import sites unchanged

Verified: tsc clean, lint 0 errors, build succeeds. No behavior change.
